simworld/requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.6
action-tutorials-py==0.20.6
actionlib-msgs==4.9.0
addict==2.4.0
aenum==3.1.16
ament-clang-format==0.12.12
ament-clang-tidy==0.12.12
ament-cmake-google-benchmark==1.3.11
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-mypy==0.12.12
ament-package==0.14.1
ament-pclint==0.12.12
ament-pep257==0.12.14
ament-pycodestyle==0.12.12
ament-pyflakes==0.12.12
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
bddl==3.5.0
blinker==1.9.0
boto3==1.39.17
botocore==1.39.17
Brotli==1.1.0
builtin-interfaces==1.2.2
camera-info-manager-py==3.1.11
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coacd==1.0.7
colorama==0.4.6
comm==0.2.2
composition-interfaces==1.2.2
conda-pack==0.8.1
ConfigArgParse==1.7.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.52.0
controller-manager-msgs==2.52.0
cryptography==41.0.7
cutie==0.3.2
cv-bridge==3.2.1
cycler==0.12.1
dash==3.2.0
debugpy==1.8.14
decorator==5.2.1
demo-nodes-py==0.20.6
diagnostic-msgs==4.9.0
distro==1.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-guard-conditions==0.15.3
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
examples-rclpy-pointcloud-publisher==0.15.3
examples-tf2-py==0.25.12
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.18.0
find-object-2d==0.7.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.5.1
future==0.18.3
gazebo-msgs==3.9.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2==2.2.1
graphviz==0.21
gymnasium==1.2.0
h11==0.16.0
h2==4.2.0
h5py==3.10.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
image-geometry==3.2.1
imageio==2.37.0
imageio-ffmpeg==0.4.9
importlib_metadata==8.7.0
iniconfig==2.1.0
interactive-markers==2.3.2
ipykernel==6.29.5
ipython==8.20.0
ipywidgets==8.1.7
isaacsim==4.1.0.0
isaacsim-app==4.1.0.0
isaacsim-asset==4.1.0.0
isaacsim-benchmark==4.1.0.0
isaacsim-code-editor==4.1.0.0
isaacsim-core==4.1.0.0
isaacsim-cortex==4.1.0.0
isaacsim-example==4.1.0.0
isaacsim-extscache-kit==4.1.0.0
isaacsim-extscache-kit-sdk==4.1.0.0
isaacsim-extscache-physics==4.1.0.0
isaacsim-gui==4.1.0.0
isaacsim-kernel==4.1.0.0
isaacsim-replicator==4.1.0.0
isaacsim-rl==4.1.0.0
isaacsim-robot==4.1.0.0
isaacsim-robot-motion==4.1.0.0
isaacsim-robot-setup==4.1.0.0
isaacsim-ros1==4.1.0.0
isaacsim-ros2==4.1.0.0
isaacsim-sensor==4.1.0.0
isaacsim-storage==4.1.0.0
isaacsim-template==4.1.0.0
isaacsim-test==4.1.0.0
isaacsim-utils==4.1.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
joint-state-publisher==2.4.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
jupytext==1.17.2
kiwisolver==1.4.8
kmeans-pytorch==0.3
laser-geometry==2.4.1
launch==1.0.11
launch-param-builder==0.1.1
launch-pytest==1.0.8
launch-ros==0.19.12
launch-testing==1.0.11
launch-testing-examples==0.15.3
launch-testing-ros==0.19.12
launch-xml==1.0.11
launch-yaml==1.0.11
lazy_loader==0.4
lifecycle-msgs==1.2.2
lifecycle-py==0.20.5
llvmlite==0.44.0
logging-demo==0.20.6
lxml==6.0.1
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
message-filters==4.3.11
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
munkres==1.1.4
narwhals==2.0.1
nav-msgs==4.9.0
nbformat==5.5.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
omegaconf==2.3.0
omnigibson==1.1.1
omniverse-kit==106.0.1.126909
ompl==1.6.0
open3d==0.16.0
openai==1.98.0
opencv-contrib-python==4.12.0.88
opencv-python==4.5.4.60
opencv-python-headless==4.12.0.88
osqp==0.6.7.post3
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
parse==1.20.2
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.6
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
progressbar==2.5
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymeshlab==2022.2.post4
pyparsing==3.2.3
pyperclip==1.9.0
pyquaternion==0.9.9
pyrealsense2==2.56.5.9235
PySide6==6.9.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-qt-binding==1.1.3
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qdldl==0.1.7.post5
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.6
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
readchar==4.2.1
referencing==0.36.2
requests==2.32.4
resource-retriever==3.1.3
retrying==1.4.1
rich==14.1.0
rmw-dds-common==1.6.0
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2cli-test-interfaces==0.18.12
ros2component==0.18.14
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2test==0.4.0
ros2topic==0.18.14
ros2trace==4.1.1
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosbag2-storage-mcap-testdata==0.15.14
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-dds-idl==0.8.1
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.26.0
rpyutils==0.2.2
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.8
rqt-gui-py==1.1.8
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.8
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
Rtree==1.2.0
s3transfer==0.13.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
setuptools==80.9.0
shape-msgs==4.9.0
shiboken6==6.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
srdfdom==2.0.7
sros2==0.10.7
stack-data==0.6.3
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
supervision==0.25.1
sympy==1.14.0
teleop-twist-keyboard==2.4.1
termcolor==2.4.0
test-launch-ros==0.19.9
test-msgs==1.2.1
test-tracetools-launch==4.1.1
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
topic-monitor==0.20.6
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
tracetools-launch==4.1.1
tracetools-read==4.1.1
tracetools-test==4.1.1
tracetools-trace==4.1.1
traitlets==5.14.3
trajectory-msgs==4.9.0
transforms3d==0.4.2
trimesh==4.0.10
triton==3.1.0
turtlesim==1.4.3
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
unicodedata2==16.0.0
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.5.0
visualization-msgs==4.9.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xacro==2.1.1
yapf==0.43.0
zipp==3.23.0
zstandard==0.23.0
